+++ db/sqlc/fake_db_adapter.go
@@ -1,0 +1,9 @@
+package db
+
+type FakeAdapter struct{}
+
+func (f *FakeAdapter) Connect(string, string) (DBTX, Querier) {
+	db := FakeDB{}
+	q := &FakeQuerier{}
+	return db, q
+}
